A whole fuckload of stuff.
Got it so the game handles input, checks to see if it's right, then shoots it off to the app, which handles it, and does necessary stuff.
Updated for flags and stuff for easier access of needed things.
Updated app to have current colony/system at start
Player empire flags
Everything